+++ sentry/src/db.rs
@@ -1,7 +1,8 @@
 use deadpool_postgres::{Manager, ManagerConfig, RecyclingMethod};
 use redis::aio::MultiplexedConnection;
-use std::env;
-use tokio_postgres::NoTls;
+use std::{env, str::FromStr};
+use tokio_postgres::{types::{accepts, FromSql, Type}, NoTls};
+
 
 use lazy_static::lazy_static;
 
@@ -51,6 +52,21 @@ lazy_static! {
 
         config
     };
+}
+
+pub struct TotalCount(pub u64);
+impl<'a> FromSql<'a> for TotalCount {
+    fn from_sql(
+        ty: &Type,
+        raw: &'a [u8],
+    ) -> Result<Self, Box<dyn std::error::Error + Sync + Send>> {
+        let str_slice = <&str as FromSql>::from_sql(ty, raw)?;
+
+        Ok(Self(u64::from_str(str_slice)?))
+    }
+
+    // Use a varchar or text, since otherwise `int8` fails deserialization
+    accepts!(VARCHAR, TEXT);
 }
 
 pub async fn redis_connection(url: &str) -> Result<MultiplexedConnection, RedisError> {
